Add the ability to load a custom welcome menu. These have been moved to modules.
